SD-1263 - Rename form to remove V2
