+++ src/pages/ipPools/EditIpPage.js
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import { Page } from '@sparkpost/matchbox';
 import _ from 'lodash';
 import { ApiErrorBanner, Loading } from 'src/components';
-import IpFormV2 from './components/IpFormV2';
+import IpForm from './components/IpFormV2';
 import { showAlert } from 'src/actions/globalAlert';
 import { listPools, updatePool } from 'src/actions/ipPools';
 import { updateSendingIp } from 'src/actions/sendingIps';
@@ -65,7 +65,7 @@ export class EditIpPage extends Component {
           />
         }
 
-        {!error && <IpFormV2 onSubmit={this.onUpdateIp}/>}
+        {!error && <IpForm onSubmit={this.onUpdateIp}/>}
       </Page>
     );
   }
