+++ src/pages/ipPools/components/IpForm.js
@@ -1,137 +1,183 @@
-import React, { Component, Fragment } from 'react';
+/* eslint-disable max-lines */
+import React, { useState } from 'react';
 import _ from 'lodash';
+import classNames from 'classnames';
 import { connect } from 'react-redux';
-import { Field, formValueSelector, getFormValues, reduxForm, submit } from 'redux-form';
+import {
+  Field,
+  formValueSelector,
+  getFormValues,
+  reduxForm,
+  submit
+} from 'redux-form';
 import { withRouter } from 'react-router-dom';
-import { Button, Panel } from '@sparkpost/matchbox';
-import { CheckboxWrapper, SelectWrapper } from 'src/components/reduxFormWrappers';
-import { ConfirmationModal, LabelledValue } from 'src/components';
+import { Button, Panel, Label } from '@sparkpost/matchbox';
+import { SelectWrapper, RadioGroup } from 'src/components/reduxFormWrappers';
+import { ConfirmationModal } from 'src/components';
+import { Card, CardTitle, CardContent } from 'src/components/card';
 import ExternalLink from 'src/components/externalLink';
-import { LINKS } from 'src/constants';
-import { IP_WARMUP_STAGES } from '../constants';
+import ButtonWrapper from 'src/components/buttonWrapper';
 import {
   getIpPools,
   selectCurrentPool,
   selectIpForCurrentPool,
   selectIpFormInitialValues
 } from 'src/selectors/ipPools';
+import { IP_WARMUP_STAGES } from '../constants';
+import styles from './IpForm.module.scss';
 
-export class IpForm extends Component {
-  state = {
-    confirmationModal: false
-  }
+const formName = 'ipForm';
 
-  submitForm = () => {
-    const { submit } = this.props;
+export const IpFormV2 = (props) => {
+  const {
+    ip,
+    pool,
+    pools,
+    isAutoWarmupEnabled,
+    handleSubmit,
+    submit,
+    submitting,
+    pristine
+  } = props;
+  const [isConfirmationModalOpen, setConfirmationModalOpen] = useState(false);
+
+  const reAssignPoolsOptions = pools.map((currentPool) => ({
+    value: currentPool.id,
+    label: (currentPool.id === pool.id) ? '-- Select a new pool --' : `${currentPool.name} (${currentPool.id})`
+  }));
+
+  const stageOptions = IP_WARMUP_STAGES.map((stage) => ({
+    label: `${stage.name} (Volume: ${stage.volume})`,
+    value: stage.id,
+    disabled: stage.id > (ip.auto_warmup_stage || 1)
+  }));
+
+  const handleConfirmClick = () => {
+    setConfirmationModalOpen(false);
+    submit(formName);
+  };
+
+  const handleUpdateSendingIPClick = () => {
+    const isEnabling = isAutoWarmupEnabled && !ip.auto_warmup_enabled;
+    const isDisabling = !isAutoWarmupEnabled && ip.auto_warmup_enabled;
+
+    if (isEnabling || isDisabling) {
+      setConfirmationModalOpen(true);
+
+      return false;
+    }
+
     return submit(formName);
   };
 
-  confirmAndSubmit = () => {
-    const { isAutoWarmupEnabled, ip } = this.props;
-    const isEnabling = isAutoWarmupEnabled && !ip.auto_warmup_enabled;
-    const isDisabling = !isAutoWarmupEnabled && ip.auto_warmup_enabled;
-    if (isEnabling || isDisabling) {
-      this.setState({ confirmationModal: true });
-      return false;
-    } else {
-      return this.submitForm();
-    }
-  }
+  return (
+    <Panel title="Sending IP Details">
+      <form onSubmit={handleSubmit}>
+        <Panel.Section>
+          <div className={styles.FieldGroup}>
+            <Label>Hostname</Label> {/* NOTE: This should not be using the HTML <label> element - just a <div> with the same styles. This is a limitation of the existing component */}
 
-  render() {
-    const { ip, pool, pools, isAutoWarmupEnabled, handleSubmit, submitting, pristine } = this.props;
-    const reAssignPoolsOptions = pools.map((currentPool) => ({
-      value: currentPool.id,
-      label: (currentPool.id === pool.id) ? '-- Select a new pool --' : `${currentPool.name} (${currentPool.id})`
-    }));
+            <p>{ip.hostname}</p>
+          </div>
 
-    const confirmationModalText = isAutoWarmupEnabled
-      ? 'Enabling Auto IP Warmup will limit the amount of traffic that you can send over this IP based on the warmup stage. Remaining traffic will be distributed amongst other IPs in the same pool or the designated overflow pool.'
-      : 'Disabling Auto IP Warmup will remove the volume restrictions from this IP. If this IP is not properly warmed, this can have negative consequences on your deliverability and sender reputation.';
+          <div className={classNames(styles.MaxWidthSM, styles.FieldGroup)}>
+            <Label id="ip_pool">Reassign IP Pool</Label>
 
-    const stageOptions = IP_WARMUP_STAGES.map((stage) => ({
-      label: `${stage.name} (Volume: ${stage.volume})`,
-      value: stage.id,
-      disabled: stage.id > (ip.auto_warmup_stage || 1)
-    }));
+            <Field
+              name="ip_pool"
+              component={SelectWrapper}
+              options={reAssignPoolsOptions}
+              disabled={submitting}
+            />
+          </div>
 
-    return (
-      <Panel>
-        <form onSubmit={handleSubmit}>
-          <Panel.Section>
-            <LabelledValue label='Hostname'>
-              <p>{ip.hostname}</p>
-            </LabelledValue>
-          </Panel.Section>
-          <Panel.Section>
-            <LabelledValue label='Reassign Pool'>
+          <fieldset className={styles.RadioGroup}>
+            <Label>Auto IP Warmup</Label> {/* NOTE: This *should* be a `<legend>` inside of a `<fieldset>` */}
+
+            <Field
+              name="auto_warmup_enabled"
+              component={RadioGroup}
+              type="radio"
+              parse={(val) => val === 'true' ? true : false}
+              value={isAutoWarmupEnabled ? 'true' : 'false'}
+              options={[
+                {
+                  label: <strong>Auto IP Warmup</strong>,
+                  value: 'true',
+                  helpText: (
+                    <div className={styles.MaxWidthMD}>
+                      <p className={styles.RadioParagraph}>Standard warmup involves gradually increasing the amount of traffic to the gold IP each day based on a pre-determined schedule. Traffic exceeding the daily limit will be directed to other IPs in the same pool and then to the overflow pool you designate.</p>
+
+                      <div className={styles.WarmupDocLink}>
+                        <ExternalLink to="https://www.sparkpost.com/docs/user-guide/automated-ip-warmup/">Warmup Documentation</ExternalLink>
+                      </div>
+                    </div>
+                  )
+                },
+                {
+                  label: <strong>Do Not Throttle for Warmup</strong>,
+                  value: 'false',
+                  helpText: <p className={classNames(styles.RadioParagraph, styles.MaxWidthMD)}>This IP will not be throttled for the purposes of IP warmup. Use this option if your IP is already warm or if you're planning to use your own warmup methods.</p>
+                }
+              ]}
+              disabled={submitting}
+            />
+          </fieldset>
+
+          {/* eslint-disable no-restricted-syntax */}
+          <div className={classNames(styles.MaxWidthMD, styles.Card)}>
+            <Card>
+              <CardTitle>Engagement Based IP Warmup</CardTitle>
+
+              <CardContent>
+                <p className={styles.CardParagraph}>This feature exists for accounts with a <abbr title="Customer Success Manager">CSM</abbr> or dedicated <abbr title="Technical Account Manager">TAM</abbr>. We will automatically bind highly engaged traffic to your cold IP to give it the best chance of warming up.</p>
+              </CardContent>
+            </Card>
+          </div>
+          {/* eslint-enable no-restricted-syntax */}
+
+          {isAutoWarmupEnabled &&
+            <div className={classNames(styles.MaxWidthSM, styles.FieldGroup)}>
               <Field
-                name='ip_pool'
+                label="Warmup Stage"
+                name='auto_warmup_stage'
                 component={SelectWrapper}
-                options={reAssignPoolsOptions}
+                options={stageOptions}
+                parse={_.toInteger}
+                helpText="You can select a previous stage but can not select an advanced stage."
                 disabled={submitting}
               />
-            </LabelledValue>
-          </Panel.Section>
-          <Panel.Section
-            actions={[{
-              color: 'orange',
-              component: ExternalLink,
-              content: 'What is Auto Warmup?',
-              to: LINKS.AUTO_IP_WARMUP_SETUP
-            }]}
-          >
-            <LabelledValue label='Auto IP Warmup'>
-              <Field
-                name="auto_warmup_enabled"
-                component={CheckboxWrapper}
-                type="checkbox"
-                label="Enable"
-                disabled={submitting}
-              />
-            </LabelledValue>
-            {isAutoWarmupEnabled &&
-            <Fragment>
-              <LabelledValue label='Warmup Stage'>
-                <Field
-                  name='auto_warmup_stage'
-                  component={SelectWrapper}
-                  options={stageOptions}
-                  parse={_.toInteger}
-                  helpText="You can select a previous stage but can not select an advanced stage."
-                  disabled={submitting}
-                />
-              </LabelledValue>
-            </Fragment>
-            }
-          </Panel.Section>
-          <Panel.Section>
-            <Button primary disabled={submitting || pristine} onClick={this.confirmAndSubmit}>
+            </div>
+          }
+        </Panel.Section>
+
+        <Panel.Section>
+          <ButtonWrapper>
+            <Button primary disabled={submitting || pristine} onClick={handleUpdateSendingIPClick}>
               {submitting ? 'Saving' : 'Update Sending IP'}
             </Button>
-          </Panel.Section>
-        </form>
+          </ButtonWrapper>
+        </Panel.Section>
+      </form>
 
-        <ConfirmationModal
-          open={this.state.confirmationModal}
-          title={`Are you sure you want to ${isAutoWarmupEnabled ? 'enable' : 'disable'} Auto IP Warmup?`}
-          content={<p>{confirmationModalText}</p>}
-          onCancel={() => this.setState({ confirmationModal: false })}
-          onConfirm={() => this.setState({ confirmationModal: false }, this.submitForm)}
-          confirmVerb={isAutoWarmupEnabled ? 'Yes, I want to turn ON Auto IP Warmup' : 'Yes, I want to turn OFF Auto IP Warmup'}
-        />
-      </Panel>
-    );
-  }
-}
-
-IpForm.defaultProps = {
-  pools: [],
-  pool: {},
-  ip: {}
+      <ConfirmationModal
+        open={isConfirmationModalOpen}
+        title={`Are you sure you want to ${isAutoWarmupEnabled ? 'enable' : 'disable'} Auto IP Warmup?`}
+        content={
+          isAutoWarmupEnabled ? (
+            <p>Enabling Auto IP Warmup will limit the amount of traffic that you can send over this IP based on the warmup stage. Remaining traffic will be distributed amongst other IPs in the same pool or the designated overflow pool.</p>
+          ) : (
+            <p>Disabling Auto IP Warmup will remove the volume restrictions from this IP. If this IP is not properly warmed, this can have negative consequences on your deliverability and sender reputation.</p>
+          )
+        }
+        onCancel={() => setConfirmationModalOpen(false)}
+        onConfirm={handleConfirmClick}
+        confirmVerb={isAutoWarmupEnabled ? 'Yes, I want to turn ON Auto IP Warmup' : 'Yes, I want to turn OFF Auto IP Warmup'}
+      />
+    </Panel>
+  );
 };
 
-const formName = 'ipForm';
 const valueSelector = formValueSelector(formName);
 const mapStateToProps = (state, props) => ({
   pool: selectCurrentPool(state, props),
@@ -147,6 +193,6 @@ const formOptions = {
   enableReinitialize: true
 };
 
-const connectedForm = withRouter(connect(mapStateToProps, { submit })(reduxForm(formOptions)(IpForm)));
+const connectedForm = withRouter(connect(mapStateToProps, { submit })(reduxForm(formOptions)(IpFormV2)));
 connectedForm.displayName = 'IpForm';
 export default connectedForm;
